--- a/core/chains/evm/logpoller/log_poller_test.go
+++ b/core/chains/evm/logpoller/log_poller_test.go
@@ -39,23 +39,17 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/utils"
 )
 
-func logRuntime(t *testing.T, start time.Time) {
+func logRuntime(t testing.TB, start time.Time) {
 	t.Log("runtime", time.Since(start))
 }
 
-func TestPopulateLoadedDB(t *testing.T) {
-	t.Skip("Only for local load testing and query analysis")
-	lggr := logger.TestLogger(t)
-	_, db := heavyweight.FullTestDBV2(t, "logs_scale", nil)
-	chainID := big.NewInt(137)
-
-	o := logpoller.NewORM(big.NewInt(137), db, lggr, pgtest.NewQConfig(true))
+func populateDatabase(t testing.TB, o *logpoller.ORM, chainID *big.Int) (common.Hash, common.Address, common.Address) {
 	event1 := EmitterABI.Events["Log1"].ID
 	address1 := common.HexToAddress("0x2ab9a2Dc53736b361b72d900CdF9F78F9406fbbb")
 	address2 := common.HexToAddress("0x6E225058950f237371261C985Db6bDe26df2200E")
+	startDate := time.Date(2010, 1, 1, 12, 12, 12, 0, time.UTC)
 
-	// We start at 1 just so block number > 0
-	for j := 1; j < 1000; j++ {
+	for j := 1; j < 100; j++ {
 		var logs []logpoller.Log
 		// Max we can insert per batch
 		for i := 0; i < 1000; i++ {
@@ -63,20 +57,57 @@ func TestPopulateLoadedDB(t *testing.T) {
 			if (i+(1000*j))%2 == 0 {
 				addr = address2
 			}
+			blockNumber := int64(i + (1000 * j))
+			blockTimestamp := startDate.Add(time.Duration(j*1000) * time.Hour)
+
 			logs = append(logs, logpoller.Log{
-				EvmChainId:  utils.NewBig(chainID),
-				LogIndex:    1,
-				BlockHash:   common.HexToHash(fmt.Sprintf("0x%d", i+(1000*j))),
-				BlockNumber: int64(i + (1000 * j)),
-				EventSig:    event1,
-				Topics:      [][]byte{event1[:], logpoller.EvmWord(uint64(i + 1000*j)).Bytes()},
-				Address:     addr,
-				TxHash:      common.HexToHash("0x1234"),
-				Data:        logpoller.EvmWord(uint64(i + 1000*j)).Bytes(),
+				EvmChainId:     utils.NewBig(chainID),
+				LogIndex:       1,
+				BlockHash:      common.HexToHash(fmt.Sprintf("0x%d", i+(1000*j))),
+				BlockNumber:    blockNumber,
+				BlockTimestamp: blockTimestamp,
+				EventSig:       event1,
+				Topics:         [][]byte{event1[:], logpoller.EvmWord(uint64(i + 1000*j)).Bytes()},
+				Address:        addr,
+				TxHash:         utils.RandomAddress().Hash(),
+				Data:           logpoller.EvmWord(uint64(i + 1000*j)).Bytes(),
+				CreatedAt:      blockTimestamp,
 			})
+
 		}
 		require.NoError(t, o.InsertLogs(logs))
+		require.NoError(t, o.InsertBlock(utils.RandomAddress().Hash(), int64((j+1)*1000-1), startDate.Add(time.Duration(j*1000)*time.Hour)))
 	}
+
+	return event1, address1, address2
+}
+
+func BenchmarkSelectLogsCreatedAfter(b *testing.B) {
+	chainId := big.NewInt(137)
+	_, db := heavyweight.FullTestDBV2(b, "logs_scale", nil)
+	o := logpoller.NewORM(chainId, db, logger.TestLogger(b), pgtest.NewQConfig(false))
+	event, address, _ := populateDatabase(b, o, chainId)
+
+	// Setting searchDate to pick around 5k logs
+	searchDate := time.Date(2020, 1, 1, 12, 12, 12, 0, time.UTC)
+
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		logs, err := o.SelectLogsCreatedAfter(address, event, searchDate, 500)
+		require.NotZero(b, len(logs))
+		require.NoError(b, err)
+	}
+}
+
+func TestPopulateLoadedDB(t *testing.T) {
+	t.Skip("Only for local load testing and query analysis")
+	_, db := heavyweight.FullTestDBV2(t, "logs_scale", nil)
+	chainID := big.NewInt(137)
+
+	o := logpoller.NewORM(big.NewInt(137), db, logger.TestLogger(t), pgtest.NewQConfig(true))
+	event1, address1, address2 := populateDatabase(t, o, chainID)
+
 	func() {
 		defer logRuntime(t, time.Now())
 		_, err1 := o.SelectLogsByBlockRangeFilter(750000, 800000, address1, event1)