--- a/core/store/migrate/migrations/0197_add_block_timestamp_index.sql
+++ b/core/store/migrate/migrations/0197_add_block_timestamp_index.sql
@@ -1,0 +1,5 @@
+-- +goose Up
+create index log_poller_blocks_by_timestamp on evm.log_poller_blocks (evm_chain_id, block_timestamp);
+
+-- +goose Down
+DROP INDEX IF EXISTS evm.log_poller_blocks_by_timestamp;