@mateusz-sekara 2023-09-21
> I'd want to see benchmarks to ensure we aren't introducing any performance regressions

Query plans look good, so I thought perf tests might not be needed. but I can do that. Btw, this query is used only by CCIP, so the impact is very localized.